Move testing package to a public place
Having it be external means it can be used in other projects. There's
really no reason to keep it internal, and it's a useful little package.
